Fixed migration bugs from moving to root

+++ db/migrate/20151015224701_create_transactions.rb
@@ -6,7 +6,7 @@ class CreateTransactions < ActiveRecord::Migration
       t.string :type
       t.string :securityPass
       t.date :expirationDate
-      t.double :price
+      t.decimal :price
 
       t.timestamps null: false
     end
